app/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "plan-search-chatbot"
version = "0.1.0"
description = "AI chatbot for Microsoft product inquiries using Azure OpenAI"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "pytest>=7.4.0",
    "pytest-repeat>=0.9.4",
    "fastapi>=0.103.0",
    "uvicorn>=0.23.0",
    "openai>=1.3.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.24.1",
    "requests==2.32.3",
    "redis>=5.0.0",
    "scrapy==2.13.0",
    "beautifulsoup4==4.12.2",
    "langchain==0.3.25",
    "azure-ai-projects==1.0.0b11",
    "azure-identity==1.23.0",
    "pytz==2025.2",
    "azure-ai-evaluation==1.8.0",
    "matplotlib==3.10.3", 
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "pytest-asyncio>=0.21.0",
    "anyio>=3.7.0",
    "pytest-tornasync>=0.6.0",
    "pytest-trio>=0.8.0",
    "pytest-twisted>=1.14.0",
    "azure-ai-evaluation[redteam]==1.8.0", 
    "ipykernel",
    
]

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[dependency-groups]
dev = [
    "pytest-asyncio>=0.26.0",
]
